Improved error handling in PlatoMain
Prior to this commit, if an exception was thrown during creation
of the PlatoApp, the exception was not caught and the MPMD run
exited without reporting and with a stack trace written to the
console.  Now the PlatoApp is created in a try-catch block, and
if an exception is thrown, the catch block calls the
Plato::Interface::Catch() function which writes the error message
to the console and registers the exception in the exceptionHandler.
Plato::Interface::registerPerformer(...) now accepts a nullptr
which causes another exception to be registered.  At the next
handleExceptions() checkin the MPMD run exits gracefully, i.e.,
there's no stack trace dumped to the console.

story: TO-898

boy scout rule: minor improvement to error reporting.