Adding support for more regressions tests.
Support option to not filter in PE. This is for cases when the filtering is happening down in PA (like for symmetry enforcement).
Support for symmetry enforcement.
Added CopyField and CopyValue operations to platomain to move data from one shared data to another. In this way we can always have Objective Value, Objective Gradient, Constraint Value, Constraint Gradient be the outcomes of the respective stages rather than having Criterion shared data being the outcome. This cleans up the code a fair amount and makes interface.xml more straight forward.